retrofit repository moved to App
