+++ app/src/main/java/com/ineedyourcode/githubapiapp/App.kt
@@ -1,0 +1,14 @@
+package com.ineedyourcode.githubapiapp
+
+import android.app.Application
+import com.ineedyourcode.githubapiapp.data.retrofit.IRetrofitGitHubRepository
+import com.ineedyourcode.githubapiapp.data.retrofit.RetrofitGitHubRepository
+
+class App : Application() {
+
+    companion object {
+        val retrofitRepository: IRetrofitGitHubRepository by lazy {
+            RetrofitGitHubRepository()
+        }
+    }
+}
