--- app/src/main/java/com/ineedyourcode/githubapiapp/domain/usecase/GetUserUsecase.kt
@@ -1,4 +1,0 @@
-package com.ineedyourcode.githubapiapp.domain.usecase
-
-interface GetUserUsecase {
-}
