--- app/src/main/java/com/ineedyourcode/githubapiapp/domain/usecase/SearchUserUsecase.kt
@@ -1,4 +1,0 @@
-package com.ineedyourcode.githubapiapp.domain.usecase
-
-interface SearchUserUsecase {
-}
