+++ app/src/main/java/com/ineedyourcode/githubapiapp/ui/screens/userdetails/UserDetailsViewModel.kt
@@ -5,16 +5,14 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.ineedyourcode.githubapiapp.data.dto.GitHubUserProfileDto
 import com.ineedyourcode.githubapiapp.data.dto.GitHubUserRepositoryDto
-import com.ineedyourcode.githubapiapp.data.dto.GitHubUserSearchResultDto
 import com.ineedyourcode.githubapiapp.data.retrofit.IRetrofitGitHubRepository
-import com.ineedyourcode.githubapiapp.data.retrofit.RetrofitGitHubRepository
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
 class UserDetailsViewModel(
+    private val retrofitRepository: IRetrofitGitHubRepository,
     private val liveData: MutableLiveData<UserDetailsState> = MutableLiveData(),
-    private val retrofitRepository: IRetrofitGitHubRepository = RetrofitGitHubRepository(),
 ) : ViewModel() {
 
     fun getLiveData(): LiveData<UserDetailsState> = liveData
