+++ app/src/main/java/com/ineedyourcode/githubapiapp/ui/screens/usersearch/UserSearchFragment.kt
@@ -3,8 +3,11 @@ package com.ineedyourcode.githubapiapp.ui.screens.usersearch
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.viewModels
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.google.android.material.snackbar.Snackbar
+import com.ineedyourcode.githubapiapp.App
 import com.ineedyourcode.githubapiapp.R
 import com.ineedyourcode.githubapiapp.databinding.FragmentUserSearchBinding
 import com.ineedyourcode.githubapiapp.ui.screens.userdetails.UserDetailsFragment
@@ -14,7 +17,15 @@ import com.ineedyourcode.githubapiapp.ui.utils.BaseFragment
 
 class UserSearchFragment :
     BaseFragment<FragmentUserSearchBinding>(FragmentUserSearchBinding::inflate) {
-    private val viewModel by viewModels<UserSearchViewModel>()
+
+    private val viewModel: UserSearchViewModel by viewModels {
+        object : ViewModelProvider.Factory {
+            @Suppress("UNCHECKED_CAST")
+            override fun <T : ViewModel> create(modelClass: Class<T>): T {
+                return UserSearchViewModel(App.retrofitRepository) as T
+            }
+        }
+    }
 
     private lateinit var userSearchAdapter: UserSearchRecyclerViewAdapter
